pyproject.toml
[tool.poetry]
name = "thirdeye-backend"
version = "0.1.0"
description = ""
authors = ["Vetrichelvan <pythonhubdev@gmail.com>"]
readme = "README.md"
packages = [
    { include = "thirdeye_backend" }
]

[tool.poetry.dependencies]
python = "^3.12"
litestar = { extras = ["sqlalchemy"], version = "^2.10.0" }
granian = { extras = ["reload"], version = "^1.5.2" }
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
litestar-granian = "^0.5.1"

[tool.poetry.group.development.dependencies]
pre-commit = "^3.7.1"
pytest = "^8.2.2"
ruff = "^0.5.2"
pytest-cov = "^5.0.0"
isort = "^5.13.2"
mypy = "^1.10.1"
pytest-env = "^1.1.3"


[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
